+++ autobahn/autobahn/wamp/router.py
@@ -24,7 +24,7 @@ from autobahn.wamp import message
 from autobahn.wamp.exception import ProtocolError
 from autobahn.wamp.broker import Broker
 from autobahn.wamp.dealer import Dealer
-from autobahn.wamp.interfaces import IRouter
+from autobahn.wamp.interfaces import IRouter, IRouterFactory
 
 
 
@@ -33,76 +33,100 @@ class Router:
    """
    Basic WAMP router.
 
-   This class implements
-
-    - :class:`autobahn.wamp.interfaces.IBroker`
-    - :class:`autobahn.wamp.interfaces.IDealer`    
+   This class implements :class:`autobahn.wamp.interfaces.IRouter`.
    """
 
-   def __init__(self, realm):
-      self._realm = realm
-      self._broker = Broker()
-      self._dealer = Dealer()
+   def __init__(self, factory, realm):
+      self.factory = factory
+      self.realm = realm
+      self._broker = Broker(realm)
+      self._dealer = Dealer(realm)
+      self._attached = 0
 
 
-   def addSession(self, session):
-      self._broker.addSession(session)
-      self._dealer.addSession(session)
-
-
-   def removeSession(self, session):
-      self._broker.removeSession(session)
-      self._dealer.removeSession(session)
-
-
-   def processMessage(self, session, msg):
+   def attach(self, session):
       """
-      Implements :func:`autobahn.wamp.interfaces.IDealer.processMessage`
+      Implements :func:`autobahn.wamp.interfaces.IRouter.attach`
+      """
+      self._broker.attach(session)
+      self._dealer.attach(session)
+      self._attached += 1
+
+
+   def detach(self, session):
+      """
+      Implements :func:`autobahn.wamp.interfaces.IRouter.detach`
+      """
+      self._broker.detach(session)
+      self._dealer.detach(session)
+      self._attached -= 1
+      if not self._attached:
+         self.factory.onLastDetach(self)
+
+
+   def process(self, session, msg):
+      """
+      Implements :func:`autobahn.wamp.interfaces.IRouter.process`
       """
       ## Broker
       ##
       if isinstance(msg, message.Publish):
-         self._broker._processPublish(session, msg)
+         self._broker.processPublish(session, msg)
 
       elif isinstance(msg, message.Subscribe):
-         self._broker._processSubscribe(session, msg)
+         self._broker.processSubscribe(session, msg)
 
       elif isinstance(msg, message.Unsubscribe):
-         self._broker._processUnsubscribe(session, msg)
+         self._broker.processUnsubscribe(session, msg)
 
       ## Dealer
       ##
       elif isinstance(msg, message.Register):
-         self._dealer._processRegister(session, msg)
+         self._dealer.processRegister(session, msg)
 
       elif isinstance(msg, message.Unregister):
-         self._dealer._processUnregister(session, msg)
+         self._dealer.processUnregister(session, msg)
 
       elif isinstance(msg, message.Call):
-         self._dealer._processCall(session, msg)
+         self._dealer.processCall(session, msg)
 
       elif isinstance(msg, message.Cancel):
-         self._dealer._processCancel(session, msg)
+         self._dealer.processCancel(session, msg)
 
       elif isinstance(msg, message.Yield):
-         self._dealer._processYield(session, msg)
+         self._dealer.processYield(session, msg)
 
       elif isinstance(msg, message.Error) and msg.request_type == message.Invocation.MESSAGE_TYPE:
-         self._dealer._processInvocationError(session, msg)
+         self._dealer.processInvocationError(session, msg)
 
       else:
          raise ProtocolError("Unexpected message {}".format(msg.__class__))
 
 
 
+@implementer(IRouterFactory)
 class RouterFactory:
+   """
+   Basic WAMP Router factory.
+
+   This class implements :class:`autobahn.wamp.interfaces.IRouterFactory`.
+   """
 
    def __init__(self):
       self._routers = {}
 
 
    def get(self, realm):
+      """
+      Implements :func:`autobahn.wamp.interfaces.IRouterFactory.get`
+      """
       if not realm in self._routers:
-         self._routers[realm] = Router(realm)
+         self._routers[realm] = Router(self, realm)
          print("Router created for realm '{}'".format(realm))
       return self._routers[realm]
+
+
+   def onLastDetach(self, router):
+      assert(router.realm in self._routers)
+      del self._routers[router.realm]
+      print("Router destroyed for realm '{}'".format(router.realm))